requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
sqlalchemy-migrate
python-dotenv
python-ldap